Move storage interface and providers to separate folder
