+++ lib/Onyx.js
@@ -1,7 +1,7 @@
 import _ from 'underscore';
 import Str from 'expensify-common/lib/str';
 import lodashMerge from 'lodash/merge';
-import Storage from './provider';
+import Storage from './storage';
 
 import {registerLogger, logInfo, logAlert} from './Logger';
 import cache from './OnyxCache';
