--- lib/__mocks__/provider.native.js
@@ -1,3 +1,0 @@
-import * as Storage from '@react-native-async-storage/async-storage/jest/async-storage-mock';
-
-export default Storage;
