+++ lib/storage/__mocks__/index.native.js
