+++ lib/storage/index.js
@@ -1,0 +1,5 @@
+import StorageProvider from './providers/AsyncStorageProvider';
+
+const instance = new StorageProvider();
+
+export default instance;
