+++ lib/storage/index.native.js
@@ -1,0 +1,5 @@
+import StorageProvider from './providers/MMKV_Provider';
+
+const instance = new StorageProvider();
+
+export default instance;
