+++ lib/storage/providers/AsyncStorageProvider.js
@@ -37,6 +37,4 @@ class AsyncStorageProvider {
     }
 }
 
-const instance = new AsyncStorageProvider();
-
-export default instance;
+export default AsyncStorageProvider;
