+++ lib/storage/providers/MMKV_Provider.js
@@ -60,6 +60,4 @@ class MMKV_Provider {
     }
 }
 
-const instance = new MMKV_Provider();
-
-export default instance;
+export default MMKV_Provider;
