+++ lib/storage/providers/ProviderInterface.js
@@ -4,6 +4,7 @@ function notImplemented() {
 
 /**
  * @interface
+ * The interface Onyx uses to communicate with the storage layer
  */
 class ProviderInterface {
     /**
